+++ dns/blocky/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.1 - 2026-09-20
+
+- Certificate for DoT/DoH can be picked from System: Trust.
+- Generated files are removed on uninstall.
+
 ## 1.0 - 2026-09-16
 
 - Initial release.